+++ vehicle-syntax/src/Vehicle/Syntax/Tensor.hs
@@ -26,6 +26,8 @@ module Vehicle.Syntax.Tensor
     toVector,
     fromVector,
     isTensorOfAll,
+    compareTensor,
+    extendTensor,
   )
 where
 
@@ -178,6 +180,11 @@ unstack xs = case shapeOf xs of
       let stride = product ds
       fmap (fromVectorSlice stride ds values) [0 .. d - 1]
 
+extendTensor :: Int -> Tensor a -> Tensor a
+extendTensor dim = \case
+  ConstantTensor shape value -> ConstantTensor (dim : shape) value
+  DenseTensor shape values -> DenseTensor (dim : shape) (Vector.concat (replicate dim values))
+
 foldMapTensor :: forall a b. (a -> b) -> (TensorShape -> [b] -> b) -> Tensor a -> b
 foldMapTensor mkValue mkVec t =
   foldMapTensorLike mkValue mkVec (shapeOf t) (toList t)
@@ -189,6 +196,9 @@ foldMapTensorLike mkValue mkVec (_ : ds) xs = do
   let inputVarIndicesChunks = chunksOf (product ds) xs
   let elems = fmap (foldMapTensorLike mkValue mkVec ds) inputVarIndicesChunks
   mkVec ds elems
+
+compareTensor :: (a -> b -> Bool) -> Tensor a -> Tensor b -> Bool
+compareTensor f t1 t2 = allTensor id $ zipWithTensor f t1 t2
 
 prettyTensor :: (a -> Doc b) -> Tensor a -> Doc b
 prettyTensor prettyElement = do