+++ vehicle/src/Vehicle/Compile/Print.hs
@@ -27,6 +27,7 @@ import Data.List.NonEmpty (NonEmpty)
 import Data.Map (Map)
 import Data.Map qualified as Map
 import Data.Text (Text)
+import Data.Tuple (swap)
 import GHC.Exts qualified as GHC (Constraint)
 import GHC.TypeLits
 import Prettyprinter (fill)
@@ -44,9 +45,10 @@ import Vehicle.Data.Code.BooleanExpr
 import Vehicle.Data.Code.LinearExpr
 import Vehicle.Data.Code.Value
 import Vehicle.Data.QuantifiedVariable (NetworkIOElementVariable, NetworkIOVariable, TensorVariable, TensorVariableLike (..), UserVariable, variableValue)
-import Vehicle.Data.Tensor (RatTensor, Tensor, prettyTensor)
+import Vehicle.Data.Tensor (RatTensor, Tensor, prettyTensor, pattern ZeroDimTensor)
 import Vehicle.Syntax.AST.Expr qualified as S
 import Vehicle.Syntax.Print
+import Vehicle.Verify.QueryFormat.Interface (QueryAssertion (..))
 import Vehicle.Verify.Specification (CompilationStep (..))
 
 --------------------------------------------------------------------------------
@@ -175,6 +177,7 @@ type family StrategyFor (tags :: Tags) a :: Strategy where
   -- To print a `Value` we need to quote it first. Note that we convert it to a `Builtin` representation immediately
   StrategyFor ('Named tags) (Value builtin `In` NamedBoundCtx) = 'QuoteValue (StrategyFor ('Named tags) (Expr Builtin `In` NamedBoundCtx))
   StrategyFor ('Unnamed tags) (Value builtin `In` ctx) = 'DescopeNaively (StrategyFor tags S.Expr)
+  StrategyFor tags (BoundEnv builtin `In` ctx) = StrategyFor tags (Value builtin `In` ctx)
   -------------------
   -- Context setup --
   -------------------
@@ -235,6 +238,8 @@ type family StrategyFor (tags :: Tags) a :: Strategy where
     'Branch
       (StrategyFor tags (TensorVariable `In` ctx))
       (StrategyFor tags (LinearExpr TensorVariable RatTensor `In` ctx))
+  StrategyFor tags (QueryAssertion variable `In` ctx) =
+    StrategyFor tags (variable `In` ctx)
   ------------
   -- Pretty --
   ------------
@@ -387,6 +392,29 @@ instance
   where
   prettyUsing (e, _ctx) = prettyUsing @rest $ fmap descopeValueNaively e
 
+instance
+  ( PrettyUsing rest (Value builtin `In` ctx),
+    PrintableBuiltin builtin
+  ) =>
+  PrettyUsing rest (BoundEnv builtin `In` ctx)
+  where
+  prettyUsing (env, ctx) = prettyEnv @rest ctx env
+
+prettyEnv :: forall rest builtin ctx a. (PrettyUsing rest (Value builtin `In` ctx)) => ctx -> BoundEnv builtin -> Doc a
+prettyEnv ctx (BoundEnv env) = prettyFlatList $ go env
+  where
+    go :: GenericBoundCtx (GenericBinder (), EnvEntry builtin) -> [Doc a]
+    go = \case
+      [] -> []
+      (binder, value) : rs -> do
+        let valueDoc = goEntry value
+        (pretty (nameOf binder) <+> "=" <+> valueDoc) : go rs
+
+    goEntry :: EnvEntry builtin -> Doc a
+    goEntry = \case
+      Bound v -> "bound" <+> prettyUsing @rest (v, ctx)
+      Unbound lv -> "unbound" <+> pretty lv
+
 -- Linear expression
 
 instance
@@ -447,6 +475,16 @@ instance
       prettyUsing @restVar (var, ctx)
         <+> "->"
         <+> prettyUsing @('Functor restVar) (childVars, ctx)
+
+instance
+  ( PrettyUsing restVar (variable `In` ctx)
+  ) =>
+  PrettyUsing restVar (QueryAssertion variable `In` ctx)
+  where
+  prettyUsing (QueryAssertion {..}, ctx) = do
+    let prettyVar u = prettyUsing @restVar (u, ctx)
+    let varCoeffs = NonEmpty.toList (fmap swap lhs)
+    prettyLinearExprLike prettyVar pretty varCoeffs (ZeroDimTensor rhs) <> pretty rel <+> "0"
 
 --------------------------------------------------------------------------------
 -- 'DescopeWithNames