+++ vehicle/src/Vehicle/Verify/QueryFormat/Interface.hs
@@ -1,10 +1,8 @@
 module Vehicle.Verify.QueryFormat.Interface where
 
-import Control.Monad.Except (MonadError)
 import Data.List.NonEmpty (NonEmpty)
 import Data.Text (Text)
-import Vehicle.Compile.Error (CompileError)
-import Vehicle.Compile.Prelude (Coefficient, ExternalOutputFormat, InputOrOutput, MonadLogger, Name, Pretty (..))
+import Vehicle.Compile.Prelude (Coefficient, ExternalOutputFormat, InputOrOutput, MonadLogger, Name)
 import Vehicle.Data.Code.BooleanExpr (ConjunctAll)
 import Vehicle.Data.Tensor (TensorIndices, TensorShape)
 import Vehicle.Verify.Core
@@ -40,13 +38,10 @@ data QueryAssertion variable = QueryAssertion
     rhs :: !Rational
   }
 
-instance (Pretty variable) => Pretty (QueryAssertion variable) where
-  pretty (QueryAssertion lhs rel rhs) = pretty lhs <> pretty rel <> pretty rhs
-
 -- | The command to format an individual query
 type CompileQuery =
   forall m.
-  (MonadLogger m, MonadError CompileError m) =>
+  (MonadLogger m) =>
   QueryAddress ->
   MetaNetwork ->
   [QueryVariable] ->